+++ aim/sdk/flush.py
@@ -1,0 +1,6 @@
+from aim.sdk.utils import get_default_session
+
+
+def flush():
+    sess = get_default_session()
+    return sess.flush()